requirements.txt
apturl==0.5.2
argcomplete==1.8.1
asgiref==3.2.10
asn1crypto==0.24.0
Babel==2.8.0
beautifulsoup4==4.6.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-ckeditor==5.9.0
django-cleanup==5.0.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
django-tinymce==3.0.2
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mysql-connector-python==2.1.6
mysqlclient==2.0.1
netifaces==0.10.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
pexpect==4.2.1
phonenumbers==8.12.6
Pillow==5.1.0
progressbar==2.3
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyMySQL==0.8.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.6.1
python-debian==0.1.32
python-gnupg==0.4.1
python3-openid==3.2.0
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sqlparse==0.3.1
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
Ubuntu-Make==18.5
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
webencodings==0.5
whitenoise==5.1.0
xkit==0.0.0
zope.interface==4.3.2
